Extract BLE client from dialog
