+++ example.html
@@ -1,6 +1,17 @@
 <html>
   <head>
     <title>Improv Example</title>
+    <style>
+      @media (prefers-color-scheme: dark) {
+        body {
+          background-color: #333;
+          color: #fff;
+        }
+        a {
+          color: #fff;
+        }
+      }
+    </style>
   </head>
   <body>
     <p>Default button:</p>
