+++ script/develop
@@ -11,7 +11,7 @@ trap "kill 0" EXIT
 # Run tsc once as rollup expects those files
 npm exec -- tsc || true
 
-npm exec -- serve &
+npm exec -- serve -p 5003 &
 npm exec -- tsc --watch &
 npm exec -- rollup -c --watch &
 wait
